refactor loading so it lives in lazy/eager, greatly simplify format
